Update CocoaPods repository (#673)
After restoring CocoaPods from cache we also need to fetch the latest
updates which were not cached yet. This is not noticeable most of the
time, but otherwise it breaks builds right after a new version of Themis
is released on CocoaPods and we can't fetch it because the repo is not
updated yet.

(cherry picked from commit 89304e51d2af7e4c0b83435573f4c58a90645284)