Move static libs and link files into dev packages
Normally the "library" packages on Debian an RHEL contain only shared
library files. Headers, static libraries, and shared library links go
into the "development" package which is necessary to build software
that uses the library (but not run it). On RHEL the static libraries
are often moved into their own "-static" package to avoid bloating the
"-devel" package (as dynamic linkage is customary).

Historically, we've been putting everything into the "library" package
but let's follow the conventions from now on.

We already recommend to install "development" packages to develop
software that uses Themis. Some language wrappers -- dynamic languages
like Python and Ruby -- will need that as they resolve Themis dynamic
library dynamically and need the symlink to be present.