Publish source and Javadoc JARs (#679)
Add tasks and artifacts for publishing JARs with source code and
generated Javadocs. These are required for admission to Bintray's
JCenter -- well-known public repository of open-source libraries.

With this, the users will be able to import Themis without adding
Cossack Labs repository. It's highly likely that they already have
JCenter added for other dependencies:

    repositories {
        jcenter()
    }

Since it's the same package, the dependency is specified as before:

    dependencies {
        implementation 'com.cossacklabs.com:themis:0.13.0'
    }

Many Bothans died to bring us this information. No, really, Gradle
documentation on this is less than stellar, and Groovy does not make
it make it easier. *sigh* Android ecosystem vOv