Use "#!/usr/bin/env bash" strictly with one argument
Apparently, Linux kernel accepts at most one argument to the binary
on the shebang line. Everything after the space is treated as a single
argument. Obviously, the user is unlikely to have an executable called
"bash -e" in their PATH, so replace all

    #!/bin/bash -e

with

    #!/usr/bin/env bash
    set -eu

instead of

    #!/usr/bin/env bash -e

which fails on Linux.

(Yeah, enable not only "-e" but "-u" as well as it's a common mistake to
use unset variables due to typos or other errors.)