Fix "libthemis" dependencies on some systems
On some systems (e.g., Ubuntu 16.04) the "libssl-dev" package depends
not only on "libsslX.Y.Z", but on other packages too ("zlib1g" in the
casse of Ubuntu). The current spell for finding out the name of the
OpenSSL library package does not take that into account, and this leads
to incorrect command line passed to FPM.

Improve the spell to react only on "libssl" packages, which there should
be only one in the dependency list.