Use "/usr/bin/env" in shebangs (#685)
* Use "/usr/bin/env" in shebangs

Instead of directly executing stuff from /bin or /usr/bin, do it via
/usr/bin/env to allow overriding the binaries via PATH. For example,
if the user has some local installation of Bash or Python which is
set as a priority in PATH, it will be used instead of the one from
hardcoded /usr/bin location.

* Use "#!/usr/bin/env bash" strictly with one argument

Apparently, Linux kernel accepts at most one argument to the binary
on the shebang line. Everything after the space is treated as a single
argument. Obviously, the user is unlikely to have an executable called
"bash -e" in their PATH, so replace all

    #!/bin/bash -e

with

    #!/usr/bin/env bash
    set -eu

instead of

    #!/usr/bin/env bash -e

which fails on Linux.

(Yeah, enable not only "-e" but "-u" as well as it's a common mistake to
use unset variables due to typos or other errors.)

* Use "-o pipefail" where pipes are used

* Handle unset variables

The "-u" option makes it an error to use unset variables. Most of the
time it is an error to do so. Where it is not an error, the ${var:-?}
syntax should be used to provide explicit default value.