Use cossacklabs/openssl-apple for Carthage
Switch from using Marcin's OpenSSL [1] to our own build of OpenSSL [2]
which provides newer OpenSSL 1.1.1g, bitcode, and other goodies.

[1]: https://github.com/krzyzanowskim/OpenSSL
[2]: https://github.com/cossacklabs/openssl-apple

The new OpenSSL is distributed as a binary-only framework. It will be
downloaded from GitHub instead of building it from source. This is not
much different from what the previous vendor did, but is more stable.

Carthage builds use the static flavor of the framework.  We have run
into issues with dynamic framworks of OpenSSL when using Carthage, but
static frameworks seems to do very good job: the resulting binaries are
smaller, apps start a bit faster, and users are freed from the hassle of
dealing with OpenSSL linkage to their app.

Note that due to the way static linkage works, we will be exporting all
OpenSSL symbols from ObjCThemis by default. In order to avoid conflicts,
export only limited subset of symbols: Objective-C classes of ObjCThemis.