Use cossacklabs/openssl-apple for CocoaPods
Switch from using Levi Groker's OpenSSL [1] to our own build [2] which
provides newer OpenSSL 1.1.1g, bitcode, and other goodies.

[1]: https://github.com/levigroker/GRKOpenSSLFramework
[2]: https://github.com/cossacklabs/openssl-apple

The new OpenSSL is distributed as a tricky pod, but for consumers like
Themis it's just a pod.

Introduce a separate subspec for the build with newer OpenSSL, and make
it the default choice. We keep the old specs around in case someone
needs them to share GRKOpenSSL or BoringSSL with other dependencies,
as it is not possible to use CLOpenSSL simultaneously with them due to
OpenSSL symbol conflicts.

The new subspec has its oddities, but it's all (un)known magic that
seems to be absolutely necessary to build Themis properly for iOS.