Improve CocoaPods version hack for linting
Turns out that the spell we used before is not effective. sed does not
understand "\s" so replace it with "[ ]" to match the leading spaces.
Now the podspec should be correctly edited for linting and validate
as if the current commit is going to be published. (The hack will be
disabled for production branches where we are linting the spec as is.)