Add arm64 to default architectures list
Make sure that ObjCThemis is compiled for arm64e architecture as noted
in Apple guidance [1]. This is important for the case when our users
would like to test teeir apps with this architecture. Currently, Xcode
does not build slices for this architecture by default, so Carthage
builds will not include it.

[1]: https://developer.apple.com/documentation/security/preparing_your_app_to_work_with_pointer_authentication