Fix NodeJS SecureSession handling nullptr (#698)
* Fix NodeJS SecureSession handling nullptr

Checking whether passed key is a private key is not enough since not all
private keys are supported. For example, `secure_session_create()` may
return NULL for RSA key. In this case, we get no exception on JS side
but as soon as we try to connect to other peer (or accept connection)
using created object, we get error. This commit makes sure we throw JS
exception in case `secure_session_create()` returns NULL.