Use Ninja for BoringSSL if available
Ninja build tool significantly reduces compilation time for BoringSSL in
default configuration. This is mostly due to the fact that it performs
parallel builds by default but there are other gains (Makefiles that
CMake generates are not that optimal). "make -j4" is also a viable
option, but it often ends up brittle.