Carthage and CocoaPods import compatibility (#703)
* Relink against "themis.framework"

Recently added "objcthemis.framework" turned out to be a mistake. It is
not possible to migrate to it because then Swift code will not be able
to find it by its module name ("themis").

Revert the changes and relink examples to "themis.framework".

* Do not link and embed "openssl.framework"

Since ObjCThemis 0.13.1, "openssl.framework" is now not necessary in
applications. It is statically linked into "themis.framework" instead.

* Use new-style Objective-C import

In order to unify import style between Carthage and CocoaPods, switch to
using module import, instead of header one.

Note, however, that this does not work in unit tests (tests/objcthemis)
for some reason, stating that "themis" module is not available there.