Exclude arm64 for iOS Simulator and macOS
As noted before several times, it is not possible to distribute several
arm64 slices in a single regular framework. It is possible with
XCFrameworks, but we'll still need to do the migration.

For now, keep arm64 disabled for iOS Simulator, disallowing its use with
Apple Silicon. CLOpenSSL dependency does not provide this slice either.

Disable arm64 for macOS as well because the current version of CLOpenSSL
does not include support for Apple Silicon.

macOS support will be enabled back with an update CLOpenSSL release, but
iOS Simulator will have to wait until the XCFramework migration.