Remove Carthage arm64 hacks from GitHub Actions
Since now arm64 is being disabled in Xcode project directly, no special
tricks are needed to disable it in Carthage builds run by CI.

However, the scripts provided for examples are kept in place for the
benefit of the users. CI tests the latest version from master, but
examples refer to 0.13.3 which still does not have arm64 exclusion.