Update supported SDKs and archs for Carthage (#733)
* Bump iOS and macOS minimum SDK versions

Just like we did for CocoaPods earlier, bump the minimum versions for
Carthage to the defaults suggested by Xcode 12.

* Exclude arm64 for iOS Simulator and macOS

As noted before several times, it is not possible to distribute several
arm64 slices in a single regular framework. It is possible with
XCFrameworks, but we'll still need to do the migration.

For now, keep arm64 disabled for iOS Simulator, disallowing its use with
Apple Silicon. CLOpenSSL dependency does not provide this slice either.

Disable arm64 for macOS as well because the current version of CLOpenSSL
does not include support for Apple Silicon.

macOS support will be enabled back with an update CLOpenSSL release, but
iOS Simulator will have to wait until the XCFramework migration.

* Remove Carthage arm64 hacks from GitHub Actions

Since now arm64 is being disabled in Xcode project directly, no special
tricks are needed to disable it in Carthage builds run by CI.

However, the scripts provided for examples are kept in place for the
benefit of the users. CI tests the latest version from master, but
examples refer to 0.13.3 which still does not have arm64 exclusion.