BoringSSL intermediate file generator
Start off by adding a utility script to generate intermediate files
required by BoringSSL build. Normally, this step is performed by the
build system configuration (e.g., CMake). It requires a number of
utilities such as Perl and Go to generate the error list and all the
assembly code. Once this is done, the intemediate files will be checked
into the repository and reused.

This script will be executed as part of Make rule for updating BoringSSL.
It is not expected to be used manually, but you can certainly do this.

Note that BoringSSL's "util/generate_build_files.py" script is pretty
picky. It expects the BoringSSL source tree to be located in "src"
directory, it works only with Python 2, and it spews out generated files
into the current directory and expects them to stay there.

Also note that we save not only the generated source files but also
the "boringssl.gypi" file. Our build system does not use GYP, but its
project file is the easiest one to parse. We'll use it later to get the
list of all BoringSSL source files to use.