New make target: update_boringssl
Add a utility target "update_boringssl" to update BoringSSL: pull the
latest source code into the submodule and run all the update scripts.

Xcode project editing scripts are run only when running on macOS since
that's required to test the updates, and likely to perform them in the
first place.