Disable CircleCI on master branch
Now that GitHub Actions have stolen the last job CircleCI had, it's no
longer necessary on the "master" branch. However, CircleCI is still
running nightly builds on the "stable" branch so we can't turn it off
just like that.

Due to the way CircleCI works, it needs to see the ".circleci/config.yml"
file on the default branch of the repo (i.e., "master"). However, we
don't need it to do anything. So put there the minimal valid config and
let CircleCI rest. It had worked hard for us all these days, after all.

Once Themis 0.14 is out, GitHub Actions configuration will be applied
to the "stable" branch, and only then CircleCI can be deconfigured and
finally retired.

(Also note that these is no replacement for "goveralls" execution that
is removed in this commit. It does not seem to be working right now, and
no one knows when exactly it stopped working. Well, we can't replace it
easily right now so GoThemis is in the same boat with other wrappers:
that is, test coverage is collected only for Themis Core. Maybe future
work on Themis 0.14 will add more coverage coverage before the release.)